重量の関係により各種からむを float 型に変更

+++ db/migrate/20200218130707_create_products.rb
@@ -4,13 +4,13 @@ class CreateProducts < ActiveRecord::Migration[6.0]
       t.boolean :lowest, default: false, null: false  # 最安値
       t.string :name              # 品名
       t.integer :quantity         # 数量
-      t.integer :weight           # 重量
+      t.float :weight             # 重量
       t.string :unit              # 単位
-      t.integer :price            # 税込金額
-      t.integer :non_taxed_price  # 税抜金額
+      t.float :price              # 税込金額
+      t.float :non_taxed_price    # 税抜金額
       t.integer :shipping_fee     # 送料
       t.integer :point            # ポイント
-      t.integer :sum              # 合計金額
+      t.float :sum                # 合計金額
       t.string :store             # ストア
       t.string :url               # URL
       t.text :description         # 備考
